requirements.txt
streamlit
tensorflow
Pillow
matplotlib
pyttsx3